v1/requirements.txt
langchain
pydantic
python-dotenv
openai
langchain-openai
langchain-community
faiss
Flask
requests